Analysis, and structure
We start saving the stocks in respective data structures. Using tuples to store the actual stock data, and using a dictionary to store all the data of a stock together. Still need to decide how I am going to store the data of all the stocks. Considering a dictionary so you can easily search through by ticker.